Add tool to track PostgreSQL Git commits and sync to SQL tables.
This adds the new Rust tool `pg-git-commit-tracker`, which syncs
meta data from the official PostgreSQL repo to a SQL table structure
consisting of the tables `commits`, `files`, and `commit_files`.

Under the hood, the `git` binary is used for efficiency; the `git2`
crate turned out to be too slow.

The SQL procedure catbench.merge_new_commits() takes care of
merging the new data, into the SQL tables, assinging bigint IDs
for the commits and files, for efficient storage and lookups.

The `catbench.commits.id` column is in git log order, that is,
id=1 is the first commit since the first git tag, REL_12_0,
which is the first commit tag we care about since that's the
oldest officially supported PostgreSQL version.