Work on caching compilations and switch compilations between runs.
This is to eliminate noise factors that might originate from
running all benchmarks at once against the same compiled version,
instead of switching between the versions for each benchmark run.

Another simpler idea which will be tested in the master branch
is to instead just stop/start PostgreSQL in between, instead
of keeping it running for all tests in a run.